api-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
email-validator
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.1.2
python-multipart